+++ packages/directives/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @viaz/directives
 
+## 0.1.5
+
+### Patch Changes
+
+- form-table deleted event add index params
+- Updated dependencies
+  - @viaz/types@0.1.5
+
+## 0.1.4
+
+### Patch Changes
+
+- fixed templte
+- Updated dependencies
+  - @viaz/types@0.1.4
+
 ## 0.1.3
 
 ### Patch Changes