+++ packages/types/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @viaz/types
 
+## 0.1.5
+
+### Patch Changes
+
+- form-table deleted event add index params
+- Updated dependencies
+  - @viaz/constants@0.1.5
+  - @viaz/enums@0.1.5
+
+## 0.1.4
+
+### Patch Changes
+
+- fixed templte
+- Updated dependencies
+  - @viaz/constants@0.1.4
+  - @viaz/enums@0.1.4
+
 ## 0.1.3
 
 ### Patch Changes